maps: migrate map0_s00/map0_s01/map0_s02
