+++ configs/maps/map0_s00.yaml
@@ -55,22 +55,15 @@ segments:
     subsegments:
       # .rodata section
       - [0x0, .rodata, map0_s00_header]
-      - [0x1050, rodata]
-      - [0x1328, .rodata, map0_s00]
-      - [0x13DC, rodata]
-      - [0x1568, .rodata, map0_s00_2]
-      - [0x15A4, rodata]
-      - [0x1910, .rodata, map0_s00_msg]
-      - [0x1F34, rodata]
+      - [0x1050, .rodata, map0_s00]
 
       # .text section
       - [0x2138, c, map0_s00]
-      - [0x9990, c, map0_s00_2]
 
       # .data section
       - [0x13D44, .data, map0_s00_anim_info]
-      - [0x1400C, data, map0_s00]
+      - [0x1400C, data]
       - [0x15D98, .data, map0_s00_header]
-      - [0x16140, .data, map0_s00_msg]
+      - [0x16140, .data, map0_s00]
       - [0x161DC, data]
   - [0x1A600]
