+++ configs/maps/map0_s02.yaml
@@ -55,19 +55,15 @@ segments:
     subsegments:
       # .rodata section
       - [0x0, .rodata, map0_s02_header]
-      - [0x1050, rodata]
-      - [0x1228, .rodata, map0_s02]
-      - [0x1278, rodata]
-      - [0x12C8, .rodata, map0_s02_msg]
-      - [0x16D0, rodata]
+      - [0x1050, .rodata, map0_s02]
 
       # .text section
       - [0x1778, c, map0_s02]
 
       # .data section
       - [0x6414, data, map0_s02_anim_info]
-      - [0x6500, data, map0_s02]
+      - [0x6500, data]
       - [0x65E4, .data, map0_s02_header]
-      - [0x67C0, .data, map0_s02_msg]
+      - [0x67C0, .data, map0_s02]
       - [0x6818, data]
   - [0xA000]
