--- src/maps/map0_s00/map0_s00_msg.c
@@ -1,29 +1,0 @@
-
-const char* g_MapMessages[] =
-{
-    #include "maps/shared/mapMsg_common.h"
-    "~J0(2.5)\tCheryl.__Where_could_you_be? ~E ",
-    "~J0(1.5)\tIt's_strange... ",
-    "~J0(2.0)\tIt's_quiet.__Too_quiet. ",
-    "~J0(3.0)\tThis_place_is ~N\n\t\t\tlike_a_ghost_town. ~E ",
-    "~J0(3.0)\tCheryl. ~E ",
-    "~J0(1.5)\tFootsteps? ~E ",
-    "~J0(2.8)\tCheryl? ~N\n\t\t\tIs_that_Cheryl!? ~E ",
-    "~J0(1.2)\tWhere_are_you_going? ~E ",
-    "~J0(1.8)\tHey_wait...stop! ~E ",
-    "\tI'd_better_follow_Cheryl... ~E ",
-    "\tThat's_strange, ~N\n\tit's_getting_darker. ~E ",
-    "\tA_dead_end? ~N\n\tWhat_the_hell!? ~E ",
-    "~J0(1.6)\tWhat_is_this? ",
-    "~J0(2.5)\tWhat's_going_on_here!? ~E ",
-    "\tWhat's_this? ~E ",
-    "\t......better_than_nothing, ~N\n\tI_guess. ~E ",
-    "\tWhat? ~N\n\tWhat's_this_doing_here? ~E ",
-    "\tBlocked_by_a_fence? ~E ",
-    "\tA_broken_wheelchair? ~N\n\tWhat's_this_doing_here? ~E ",
-    "\tWhat_the..? ~N\n\tWhat_is_it? ~E ",
-    "\tThe_street_is_blocked.",
-    "\tCan't_get_out_of_town ~N\n\tthrough_here. ~E ",
-    "\tMy_car... ~N\n\ttoo_banged_up_to_drive. ",
-    "\tWhere's_Cheryl? ~N\n\tHope_she's_safe. ~E ",
-};
