--- src/maps/map0_s01/map0_s01_msg.c
@@ -1,72 +1,0 @@
-#include "types.h"
-
-const char* g_MapMessages[] =
-{
-    #include "maps/shared/mapMsg_common.h"
-    "~J0(2.0)\tWas_I_dreaming?  ~E ",
-    "~J0(2.5)\tHow_do_you_feel? ~E ",
-    "~J0(6.0)\tLike_I've_been ~N\n\t\t\trun_over_by_a_truck, ~N\n\t\t\tbut_I'm_all_right,_I_guess. ~E ",
-    "~J0(1.5)\tGlad_to_hear_it.",
-    "~J0(5.0)\tYou_from_around_here? ~N\n\t\t\tWhy_don't_you ~N\n\t\t\ttell_me_what_happened? ",
-    "~J0(3.0)\tWait_a_second. ~N\n\t\t\tI'm_just_a_tourist. ",
-    "~J0(5.2)\tI_came_here_for_a_vacation. ~N\n\t\t\tI_just_got_here... ",
-    "~J0(4.2)\tI_don't_know_what_happened. ",
-    "~J0(4.0)\tI'd_like_to_find_out_myself. ",
-    "~J0(2.0)\tUh-huh? ~E ",
-    "~J0(3.0)\tHave_you_seen_a_little_girl? ",
-    "~J0(4.0)\tJust_turned_seven_last_month... ~N\n\t\t\tShort,_black_hair... ",
-    "~J0(1.5)\tMy_daughter. ",
-    "~J0(1.0)\tSorry. ",
-    "~J0(3.0)\tThe_only_person_I've_seen ~N\n\t\t\tin_this_town_is_you. ",
-    "~J0(2.5)\tWhere_is_everybody? ",
-    "~J0(3.5)\tI'd_tell_you_if_I_knew,~N\n\t\t\tbelieve_me. ",
-    "~J0(5.5)\tBut,_from_what_I_can_tell, ~N\n\t\t\tsomething_bizarre_is_going_on. ",
-    "~J0(2.5)\tThat's_all_I_know. ",
-    "~J0(2.0)\tHmmm... ~E ",
-    "~J0(2.0)\tWhat's_your_name? ",
-    "~J0(1.0)\tHarry... ",
-    "~J0(2.0)\tHarry_Mason. ",
-    "~J0(2.0)\tCybil_Bennett. ",
-    "~J0(4.0)\tI'm_a_police_officer ~N\n\t\t\tfrom_Brahms, ~N\n\t\t\tthe_next_town_over. ",
-    "~J0(3.0)\tThe_phones_are_all_dead, ~N\n\t\t\tand_the_radio,_too. ",
-    "~J0(4.0)\tI'm_going_back_to_call ~N\n\t\t\tin_some_reinforcements. ",
-    "~J0(1.5)\tHmph. ~E ",
-    "~D\t~J0(4.0)\tHold_it! ~N\n\t\t\tWhere_do_you_think ~N\n\t\t\tyou're_going? ~E ",
-    "~J0(4.0)\tMy_daughter... ~N\n\t\t\tI've_gotta_find_her! ~E ",
-    "~D\t~J0(1.5)\tNo_way! ~E ",
-    "~J0(3.0)\tIt's_dangerous_out_there. ",
-    "~J0(5.0)\tIn_that_case, ~N\n\t\t\tI_need_to_find_her_now! ~N\n\t\t\tCheryl's_my_little_girl. ",
-    "~J0(3.0)\tI_can't_just_leave_her_out ~N\n\t\t\tthere_by_herself. ~E ",
-    "~J0(2.5)\tHave_you_got_a_gun? ",
-    "~J0(3.0)\tUmmm..._no. ~E ",
-    "~J0(5.0)\tTake_this. ~N\n\t\t\tAnd_hope_you_don't ~N\n\t\t\thave_to_use_it. ",
-    "~J0(1.5)\tNow_listen_to_me, ",
-    "~J0(4.5)\tbefore_you_pull_the_trigger, ~N\n\t\t\tknow_who_you're_shooting. ",
-    "~J0(3.0)\tAnd_don't_do_it ~N\n\t\t\tunless_you_have_to. ~E ",
-    "~J0(3.0)\tAnd_don't_go_blasting_me ~N\n\t\t\tby_mistake. ",
-    "~J0(2.5)\tGot_it? ~E ",
-    "~J0(2.5)\tYeah._Thanks. ~E ",
-    "~J0(5.0)\tYou'd_do_best_to_stay_nearby. ~N\n\t\t\tI'll_be_back_with ~N\n\t\t\thelp_as_quick_as_I_can. ~E ",
-    "~H There_is_a_ ~C5 Residential_area_map ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tThere_is_a_ ~C2 Kitchen_knife ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tThere_is_a_ ~C2 Flashlight ~C7 . ~N\n\tTake_it? ~S4 ",
-    "~J0(1.1)\tWhat's_that? ~E ",
-    "~J0(4.5)\tHuh?_Radio? ~N\n\t\t\tWhat's_going_on_with_that_radio? ~E ",
-    "\tThere_is_a_ ~C2 Pocket_radio ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tA_radio... is_it_broken? ~N\n\tCan't_hear_anything. ~E ",
-    " ",
-    "\tMaybe_I_should_check_for ~N\n\tsomething_useful_before_leaving. ~E ",
-    "~J0(4.3)\tThis_is_not_a_dream! ~N\n\tWhat's_happening_to_this_place!? ~E ",
-    "\tI'll_worry_about_the_radio_later, ~N\n\tI_have_more_important_stuff ~N\n\tto_do. ~E ",
-    "\tI've_got_to_beat_that_monster! ~E ",
-    "\tThe_radio_may_be ~N\n\tuseful_for_something. ",
-    "\tI_better_take_it_with_me. ~E ",
-    "\tPinball_machine. ~N\n\tNothing_unusual. ~E ",
-    "\tCan't_turn_on_the_TV. ~N\n\tNo_power? ~E ",
-    "\tI_have_more_important_stuff_to_do. ~E ",
-};
-
-const VECTOR3 D_800CC8FC = {0x00000CCC, 0xFFFFECCD, 0x0010D800};
-const char D_800CC908[] = "HERO";
-const char D_800CC910[] = "SIBYL";
-
