+++ src/maps/map0_s02/map0_s02.c
@@ -132,6 +132,18 @@ INCLUDE_ASM("asm/maps/map0_s02/nonmatchings/map0_s02", func_800CE934);
 
 void Gfx_LoadingScreen_StageString(void) {}
 
+const char* g_MapMessages[] =
+{
+    #include "maps/shared/mapMsg_common.h"
+    "\tThere_is_a_ ~C2 Chain_saw ~C7 . ~N\n\tTake_it? ~S4 ",
+    "\tSome_oil_cans_are_on_the_shelf. ~E ",
+    "\tA_car_engine. ~N\n\tNo_problem_there. ~E ",
+    "\tThere's_a_TV_set. ~N\n\tIt_doesn't_turn_on. ~E ",
+    "\tThere_is_a_ ~C2 Gasoline_tank ~C7 . ~N\n\tTake_it? ~S4 ",
+    "\tThere_is_a_ ~C2 Katana ~C7 . ~N\n\tTake_it? ~S4 ",
+    "\tThere_is_a_ ~C2 Channeling_stone ~C7 . ~N\n\tTake_it? ~S4 ",
+};
+
 // TODO: .rodata migration.
 #ifdef NON_MATCHING
 void Event_HealthOrAmmoItemTake(void) // 0x800CEBC8
@@ -205,3 +217,5 @@ void Event_ChannelingStoneItemTake(void) // 0x800CECDC
 INCLUDE_ASM("asm/maps/map0_s02/nonmatchings/map0_s02", func_800CED08);
 
 INCLUDE_ASM("asm/maps/map0_s02/nonmatchings/map0_s02", func_800CEFD4);
+
+INCLUDE_RODATA("asm/maps/map0_s02/nonmatchings/map0_s02", D_800CACE8);
