--- src/maps/map0_s02/map0_s02_msg.c
@@ -1,12 +1,0 @@
-
-const char* g_MapMessages[] =
-{
-    #include "maps/shared/mapMsg_common.h"
-    "\tThere_is_a_ ~C2 Chain_saw ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tSome_oil_cans_are_on_the_shelf. ~E ",
-    "\tA_car_engine. ~N\n\tNo_problem_there. ~E ",
-    "\tThere's_a_TV_set. ~N\n\tIt_doesn't_turn_on. ~E ",
-    "\tThere_is_a_ ~C2 Gasoline_tank ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tThere_is_a_ ~C2 Katana ~C7 . ~N\n\tTake_it? ~S4 ",
-    "\tThere_is_a_ ~C2 Channeling_stone ~C7 . ~N\n\tTake_it? ~S4 ",
-};
